Expand the set of characters allowed in words
Before this commit, words could only contain alphabetic and numeric characters.
This is unduly limiting; for example, combining marks (like accents) are not included.
This commit expands the set of allowed characters based on the recommendations of
UAX 31 <https://www.unicode.org/reports/tr31/>, with a bias toward allowing more characters.
In addition, unassigned and private use characters are assumed to be allowed in words
(ensuring that case conversion will pass them through unchanged).

One tricky case is the the asterisk (*), which is a Other_Punctuation character,
but becomes an emoji when followed bu U+FE0F (*️). Becasue it can be an emoji,
and some programming languages allow emoji in identifiers, it gets included
in the set of characters allowed in words. Special handling may be warranted here.

This change means we can no longer rely only on the Unicode data tables
shipped with the standard library.
A new `tables` binary crate is in charge of generating the tables we need
(which consume 3672 bytes of data).