Properly handle final sigma lowercase
The rules for detemining how a Greek sigma should be lowercased
are more complex than just "is it the last charater in the word."
The full rule is defined in the Unicode Standard under "Final_Sigma":
<https://www.unicode.org/versions/Unicode15.1.0/ch03.pdf#G54277>

Implementing the rules ourselves would require shipping additional static data for the `Case_Ignorable` Unicode property.
So this commit instead uses the standard library implementation,
which unfortunately costs an extra allocation.